+++ src/components/Sidebar/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-restricted-exports
+export {default} from './Connector';
